saas_app/requirements.txt
amqp==5.3.1
arabic-reshaper==3.0.0
asgiref==3.11.0
asn1crypto==1.5.1
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cssselect2==0.8.0
cycler==0.12.1
Django==6.0.1
django-cors-headers==4.9.0
django-crontab==0.7.1
djangorestframework==3.16.1
flower==2.0.1
fonttools==4.61.1
freetype-py==2.5.1
html5lib==1.1
humanize==4.15.0
idna==3.11
iniconfig==2.3.0
kiwisolver==1.4.9
kombu==5.6.2
lxml==6.0.2
matplotlib==3.10.8
numpy==2.4.1
oscrypto==1.3.0
packaging==26.0
pillow==12.1.0
pluggy==1.6.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycairo==1.29.0
pycparser==3.0
Pygments==2.19.2
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
pyparsing==3.3.2
pypdf==6.6.2
pytest==9.0.2
pytest-django==4.11.1
python-bidi==0.6.7
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
reportlab==4.4.9
requests==2.32.5
rlPyCairo==0.4.0
six==1.17.0
sqlparse==0.5.5
svglib==1.6.0
tinycss2==1.5.1
tornado==6.5.4
tzdata==2025.3
tzlocal==5.3.1
uritools==6.0.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.5.3
webencodings==0.5.1
xhtml2pdf==0.2.17
